requirements (1).txt
pandas
numpy
requests
scipy
PyYAML
streamlit
firebase-admin